+++ podcasts/Analytics/AnalyticsEvent.swift
@@ -599,6 +599,7 @@ enum AnalyticsEvent: String {
     case searchFailed
     case searchResultTapped
     case searchListShown
+    case searchCleared
 
     // MARK: - Chromecast
 
@@ -725,4 +726,18 @@ enum AnalyticsEvent: String {
     case shareScreenPlayTapped
     case shareScreenPauseTapped
     case shareScreenClipShared
+
+    // MARK: - Referrals
+
+    case referralTooltipShow
+    case referralTooltipTapped
+    case referralShareScreenShown
+    case referralPassShared
+    case referralClaimScreenShown
+    case referralActivateTapped
+    case referralNotNowTapped
+    case referralUsedScreenShown
+    case referralPassBannerShown
+    case referralPurchaseShown
+    case referralPurchaseSuccess
 }