+++ podcasts/ListeningHistoryViewController.swift
@@ -1,14 +1,37 @@
 import DifferenceKit
 import PocketCastsDataModel
 import PocketCastsServer
+import PocketCastsUtils
 import UIKit
 
 class ListeningHistoryViewController: PCViewController {
     var episodes = [ArraySection<String, ListEpisode>]()
+    var tempEpisodes = [ArraySection<String, ListEpisode>]()
     private let operationQueue = OperationQueue()
     var cellHeights: [IndexPath: CGFloat] = [:]
 
     private let episodesDataManager = EpisodesDataManager()
+    private var searchController: PCSearchBarController?
+
+    @IBOutlet weak var emptyStateView: ThemeableView! {
+        didSet {
+            emptyStateView.isHidden = true
+        }
+    }
+
+    @IBOutlet weak var emptyStateTitle: ThemeableLabel! {
+        didSet {
+            emptyStateTitle.style = .primaryText01
+            emptyStateTitle.text = L10n.listeningHistorySearchNoEpisodesTitle
+        }
+    }
+
+    @IBOutlet weak var emptyStateText: ThemeableLabel! {
+        didSet {
+            emptyStateText.style = .primaryText02
+            emptyStateText.text = L10n.listeningHistorySearchNoEpisodesText
+        }
+    }
 
     @IBOutlet var listeningHistoryTable: UITableView! {
         didSet {
@@ -65,6 +88,10 @@ class ListeningHistoryViewController: PCViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+
+        if FeatureFlag.listeningHistorySearch.enabled {
+            setupSearchController()
+        }
 
         operationQueue.maxConcurrentOperationCount = 1
         title = L10n.listeningHistory
@@ -179,5 +206,78 @@ class ListeningHistoryViewController: PCViewController {
 extension ListeningHistoryViewController: AnalyticsSourceProvider {
     var analyticsSource: AnalyticsSource {
         .listeningHistory
+    }
+}
+
+// MARK: - Analytics
+
+extension ListeningHistoryViewController: PCSearchBarDelegate {
+    func searchDidBegin() {
+        tempEpisodes = episodes
+    }
+
+    func searchDidEnd() {
+        listeningHistoryTable.isHidden = tempEpisodes.isEmpty
+        emptyStateView.isHidden = true
+        episodes = tempEpisodes
+        listeningHistoryTable.reloadData()
+        tempEpisodes.removeAll()
+    }
+
+    func searchWasCleared() {
+        Analytics.track(.searchCleared, source: analyticsSource)
+
+        listeningHistoryTable.isHidden = tempEpisodes.isEmpty
+        emptyStateView.isHidden = true
+        episodes = tempEpisodes
+        listeningHistoryTable.reloadData()
+    }
+
+    func searchTermChanged(_ searchTerm: String) { }
+
+    func performSearch(searchTerm: String, triggeredByTimer: Bool, completion: @escaping (() -> Void)) {
+        Analytics.track(.searchPerformed, source: analyticsSource)
+
+        let oldData = episodes
+        let newData = episodesDataManager.searchEpisodes(for: searchTerm)
+
+        listeningHistoryTable.isHidden = newData.isEmpty
+        emptyStateView.isHidden = !newData.isEmpty
+
+        let changeSet = StagedChangeset(source: oldData, target: newData)
+        self.listeningHistoryTable.reload(using: changeSet, with: .none, setData: { data in
+            self.episodes = data
+        })
+        completion()
+    }
+
+    private func setupSearchController() {
+        searchController = PCSearchBarController()
+        searchController?.searchDebounce = 0.2
+
+        guard let searchController else {
+            return
+        }
+
+        searchController.view.translatesAutoresizingMaskIntoConstraints = false
+        addChild(searchController)
+        view.addSubview(searchController.view)
+        searchController.didMove(toParent: self)
+
+        let topAnchor = searchController.view.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor)
+        NSLayoutConstraint.activate([
+            searchController.view.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            searchController.view.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            searchController.view.heightAnchor.constraint(equalToConstant: PCSearchBarController.defaultHeight),
+            topAnchor
+        ])
+
+        searchController.placeholderText = L10n.search
+        searchController.searchControllerTopConstant = topAnchor
+        searchController.setupScrollView(listeningHistoryTable, hideSearchInitially: false)
+        searchController.searchDebounce = Settings.podcastSearchDebounceTime()
+        searchController.searchDelegate = self
+
+        listeningHistoryTable.verticalScrollIndicatorInsets.top = PCSearchBarController.defaultHeight
     }
 }