+++ podcasts/NavigationManager.swift
@@ -55,6 +55,8 @@ class NavigationManager {
     static let settingsAppearanceKey = "appearancePage"
     static let settingsProfileKey = "profilePage"
     static let settingsHeadphoneKey = "headphoneSettings"
+    static let settingsRedeemGuestPassKey = "redeemGuestPassPage"
+    static let redeemGuestPassURLKey = "redeemGuestPassURLKey"
 
     static let endOfYearStories = "endOfYearStories"
     static let onboardingFlow = "onboardingFlow"
@@ -170,6 +172,12 @@ class NavigationManager {
         }
         else if place == NavigationManager.settingsHeadphoneKey {
             mainController?.showHeadphoneSettings()
+        }
+        else if place == NavigationManager.settingsRedeemGuestPassKey {
+            guard let data = data, let url = data[NavigationManager.redeemGuestPassURLKey] as? URL else {
+                return
+            }
+            mainController?.showRedeemGuestPass(url: url)
         }
         else if place == NavigationManager.showPromotionPageKey {
             var promoCode: String?