+++ podcasts/Referrals/ReferralSendPassView.swift
@@ -1,94 +1,134 @@
 import SwiftUI
+import PocketCastsServer
 
-class ReferralSendPassModel {
+@MainActor
+class ReferralSendPassModel: ObservableObject {
+
+    enum LoadingState {
+        case idle
+        case loading
+        case failed
+        case loaded
+    }
+
     let offerInfo: ReferralsOfferInfo
-    let numberOfPasses: Int
+    @Published var referralURL: URL?
+
     var onShareGuestPassTap: (() -> ())?
     var onCloseTap: (() -> ())?
 
-    init(offerInfo: ReferralsOfferInfo, numberOfPasses: Int = 3, onShareGuestPassTap: (() -> ())? = nil) {
+    @Published var state: LoadingState = .idle
+
+    init(offerInfo: ReferralsOfferInfo, onShareGuestPassTap: (() -> ())? = nil, onCloseTap: (() -> ())? = nil) {
         self.offerInfo = offerInfo
-        self.numberOfPasses = numberOfPasses
-        self.onShareGuestPassTap = nil
+        self.onShareGuestPassTap = onShareGuestPassTap
+        self.onCloseTap = onCloseTap
     }
 
     var title: String {
-        if numberOfPasses > 0 {
-            L10n.referralsTipMessage(offerInfo.localizedOfferDuration)
-        } else {
-            L10n.referralsShareNoGuestPassTitle
-        }
-    }
-
-    var message: String {
-        if numberOfPasses > 0 {
-            L10n.referralsTipTitle(numberOfPasses)
-        } else {
-            L10n.referralsShareNoGuestPassMessage
-        }
+        L10n.referralsTipMessage(offerInfo.localizedOfferDurationNoun)
     }
 
     var buttonTitle: String {
-        if numberOfPasses > 0 {
-            return L10n.referralsShareGuestPass
+        L10n.referralsShareGuestPass
+    }
+
+    var shareText: String {
+        L10n.referralsSharePassMessage(self.offerInfo.localizedOfferDurationAdjective)
+    }
+
+    var shareSubject: String {
+        L10n.referralsSharePassSubject(self.offerInfo.localizedOfferDurationAdjective)
+    }
+
+    func loadData() async {
+        state = .loading
+        let code = await ApiServerHandler.shared.getReferralCode()
+        if let code, let referralURL = URL(string: code.url) {
+            self.referralURL = referralURL
+            state = .loaded
         } else {
-            return L10n.gotIt
+            state = .failed
         }
     }
 }
 
 struct ReferralSendPassView: View {
-    let viewModel: ReferralSendPassModel
+    @StateObject var viewModel: ReferralSendPassModel
 
-    var body: some View {
+    @State var showShareView: Bool = false
+
+    @ViewBuilder
+    var loadingView: some View {
         VStack {
-            ModalCloseButton(background: Color.gray.opacity(0.2), foreground: Color.white.opacity(0.5), action: { viewModel.onCloseTap?() })
-            VStack(spacing: Constants.verticalSpacing) {
-                SubscriptionBadge(tier: .plus, displayMode: .gradient, foregroundColor: .black)
-                Text(viewModel.title)
-                    .font(size: 31, style: .title, weight: .bold)
-                    .multilineTextAlignment(.center)
-                    .foregroundColor(.white)
-                Text(viewModel.message)
-                    .font(size: 14, style: .body, weight: .medium)
-                    .multilineTextAlignment(.center)
-                    .foregroundColor(.white)
-                ZStack {
-                    ForEach(0..<viewModel.numberOfPasses, id: \.self) { i in
-                        ReferralCardView(offerDuration: viewModel.offerInfo.localizedOfferDuration)
-                            .frame(width: Constants.defaultCardSize.width - (CGFloat(viewModel.numberOfPasses-1-i) * Constants.cardInset.width), height: Constants.defaultCardSize.height)
-                            .offset(CGSize(width: 0, height: CGFloat(viewModel.numberOfPasses * i) * Constants.cardInset.height))
-                    }
-                }
+            Spacer()
+            HStack {
+                Spacer()
+                ProgressView().tint(.white)
+                Spacer()
             }
             Spacer()
-            Button(viewModel.buttonTitle) {
-                if viewModel.numberOfPasses > 0 {
-                    viewModel.onShareGuestPassTap?()
-                } else {
-                    viewModel.onCloseTap?()
-                }
-            }.buttonStyle(PlusGradientFilledButtonStyle(isLoading: false, plan: .plus))
         }
-        .padding()
         .background(.black)
+    }
+
+    var body: some View {
+        switch viewModel.state {
+        case .idle:
+            Color.black.task {
+                await viewModel.loadData()
+            }
+        case .loading:
+            loadingView
+        case .failed:
+            ReferralsMessageView(title: L10n.referralsNotAvailableToSend, detail: L10n.pleaseTryAgainLater) {
+                viewModel.onCloseTap?()
+            }
+        case .loaded:
+            VStack {
+                HStack {
+                    Button(action: {
+                        viewModel.onCloseTap?()
+                    }, label: {
+                        Image("close").foregroundColor(Color.white)
+                    })
+                    Spacer()
+                }
+                VStack(spacing: Constants.verticalSpacing) {
+                    SubscriptionBadge(tier: .plus, displayMode: .gradient, foregroundColor: .black)
+                    Text(viewModel.title)
+                        .font(size: 31, style: .title, weight: .bold)
+                        .multilineTextAlignment(.center)
+                        .foregroundColor(.white)
+                    ZStack {
+                        ForEach(0..<Constants.numberOfPasses, id: \.self) { i in
+                            ReferralCardView(offerDuration: viewModel.offerInfo.localizedOfferDurationAdjective)
+                                .frame(width: Constants.defaultCardSize.width - (CGFloat(Constants.numberOfPasses-1-i) * Constants.cardInset.width), height: Constants.defaultCardSize.height)
+                                .offset(CGSize(width: 0, height: CGFloat(Constants.numberOfPasses * i) * Constants.cardInset.height))
+                        }
+                    }
+                }
+                Spacer()
+                Button(viewModel.buttonTitle) {
+                    viewModel.onShareGuestPassTap?()
+                }
+                .buttonStyle(PlusGradientFilledButtonStyle(isLoading: false, plan: .plus))
+            }
+            .padding()
+            .background(.black)
+        }
     }
 
     enum Constants {
         static let verticalSpacing = CGFloat(24)
         static let defaultCardSize = ReferralCardView.Constants.defaultCardSize
         static let cardInset = CGSize(width: 40, height: 5)
+        static let numberOfPasses: Int = 3
     }
 }
 
 #Preview("With Passes") {
     Group {
         ReferralSendPassView(viewModel: ReferralSendPassModel(offerInfo: ReferralsOfferInfoMock()))
-    }
-}
-
-#Preview("Without Passes") {
-    Group {
-        ReferralSendPassView(viewModel: ReferralSendPassModel(offerInfo: ReferralsOfferInfoMock(), numberOfPasses: 0))
     }
 }