+++ podcasts/Sharing/Clip/TrimSelectionView.swift
@@ -2,7 +2,10 @@ import SwiftUI
 
 struct TrimSelectionView: View {
     @Binding var leading: CGFloat
+    @Binding var leadingA11yValue: String
     @Binding var trailing: CGFloat
+    @Binding var trailingA11yValue: String
+
     let handleWidth: CGFloat
     let indicatorWidth: CGFloat
 
@@ -21,7 +24,31 @@ struct TrimSelectionView: View {
                 .frame(width: (trailing - leading + indicatorWidth) + (Constants.borderWidth * 2))
                 .offset(x: leading - Constants.borderWidth)
             TrimHandle(position: $leading, side: .leading, width: handleWidth, onChanged: { changed($0, .leading) })
+                .accessibilityLabel(L10n.clipsEndTimeAccessibilityLabel)
+                .accessibilityValue(leadingA11yValue)
+                .accessibilityAdjustableAction { direction in
+                    switch direction {
+                    case .increment:
+                        changed(leading - (handleWidth/2) + 2, .leading)
+                    case .decrement:
+                        changed(leading - (handleWidth/2) - 2, .leading)
+                    @unknown default:
+                        break
+                    }
+                }
             TrimHandle(position: $trailing, side: .trailing, width: handleWidth, onChanged: { changed( $0, .trailing) })
+                .accessibilityLabel(L10n.clipsEndTimeAccessibilityLabel)
+                .accessibilityValue(trailingA11yValue)
+                .accessibilityAdjustableAction { direction in
+                    switch direction {
+                    case .increment:
+                        changed(trailing + (handleWidth/2) + 2, .trailing)
+                    case .decrement:
+                        changed(trailing + (handleWidth/2) - 2, .trailing)
+                    @unknown default:
+                        break
+                    }
+                }
         }
     }
 }